Create WaitAndTimeoutKeywords
