+++ Mainframe3270/keywords/__init__.py
@@ -1,2 +1,4 @@
 from Mainframe3270.keywords.assertions import AssertionKeywords  # noqa: F401
 from Mainframe3270.keywords.connection import ConnectionKeywords  # noqa: F401
+from Mainframe3270.keywords.wait_and_timeout import \
+    WaitAndTimeoutKeywords  # noqa: F401
